fix all the unused places
