+++ util/chunk/disk.go
@@ -99,34 +99,6 @@ func (l *ListInDisk) GetDiskTracker() *disk.Tracker {
 	return l.diskTracker
 }
 
-// flush empties the write buffer, please call flush before read!
-func (l *ListInDisk) flush() (err error) {
-	// buffered is not zero only after Add and before GetRow, after the first flush, buffered will always be zero,
-	// hence we use a RWLock to allow quicker quit.
-	l.bufFlushMutex.RLock()
-	checksumWriter := l.w
-	l.bufFlushMutex.RUnlock()
-	if checksumWriter == nil {
-		return nil
-	}
-	l.bufFlushMutex.Lock()
-	defer l.bufFlushMutex.Unlock()
-	if l.w != nil {
-		err = l.w.Close()
-		if err != nil {
-			return
-		}
-		l.w = nil
-		// the l.disk is the underlying object of the l.w, it will be closed
-		// after calling l.w.Close, we need to reopen it before reading rows.
-		l.disk, err = os.Open(l.disk.Name())
-		if err != nil {
-			return errors2.Trace(err)
-		}
-	}
-	return
-}
-
 // Add adds a chunk to the ListInDisk. Caller must make sure the input chk
 // is not empty and not used any more and has the same field types.
 // Warning: do not mix Add and GetRow (always use GetRow after you have added all the chunks), and do not use Add concurrently.
